+++ test/instream/response_test.exs
@@ -28,7 +28,7 @@ defmodule Instream.ResponseTest do
   test "raw json error response" do
     error = "text"
     response = {500, [{"Content-Type", "application/json"}], error}
-    parse_opts = [json_decoder: Poison, result_as: :raw]
+    parse_opts = [json_decoder: Jason, result_as: :raw]
 
     assert error == Response.maybe_parse(response, parse_opts)
   end
@@ -36,14 +36,14 @@ defmodule Instream.ResponseTest do
   test "raw non-json error response" do
     error = "text"
     response = {500, [], error}
-    parse_opts = [json_decoder: Poison, result_as: :raw]
+    parse_opts = [json_decoder: Jason, result_as: :raw]
 
     assert error == Response.maybe_parse(response, parse_opts)
   end
 
   test "regular non-json response" do
     response = "text"
-    parse_opts = [json_decoder: Poison]
+    parse_opts = [json_decoder: Jason]
 
     assert response == Response.maybe_parse({200, [], response}, parse_opts)
   end